Make the Required Changes
